+++ docs/en/changes/changes.md
@@ -45,6 +45,7 @@
   instance.
 * Fix Zipkin trace query the max size of spans.
 * Add `tls` and `https` component IDs for Network Profiling.
+* Support Elasticsearch column alias for the compatibility between storage logicSharding model and no-logicSharding model.
 
 #### UI